move Header and files in its own folder
