+++ packages/create-svelte/templates/default/src/lib/header/Header.svelte
